requirements.txt
google-generativeai
python-telegram-bot
flask
pydub
requests
flask-cors